% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle.r
\name{perm_test}
\alias{perm_test}
\title{Permutation test}
\usage{
perm_test(data, statistic, R, ..., ncores = 1L, cl = NULL)
}
\arguments{
\item{data}{A data set with at least 2 columns. The first columns holds times,
and the second column individual ids.}

\item{statistic}{A function that returns a vector. The statistic to compute}

\item{R}{Integer scalar. Number of replications.}

\item{...}{Further arguments passed to \code{statistic}}

\item{ncores}{integer scalar. Number of cores to use.}

\item{cl}{An object of class cluster.}
}
\description{
Permutation test
}
\details{
This function relies on the \link{shuffle_bites}.
}
