% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{plot.bite_dyad}
\alias{plot.bite_dyad}
\title{Plotting method}
\usage{
\method{plot}{bite_dyad}(x, y = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{bite_dyad}}

\item{y}{Ignored.}

\item{...}{Ignored.}
}
\description{
Plotting method
}
