% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{simulate_dyad}
\alias{simulate_dyad}
\title{Simulate a dyad}
\usage{
simulate_dyad(rates = c(10, 15), mimicy = c(0.1, 0.5),
  last.bite = stats::runif(1, 1, 2))
}
\arguments{
\item{rates}{Numeric vector. Vector of rates}

\item{mimicy}{Numeric vector. proportion of mimicry (likelihood).}

\item{last.bite}{Numeric scalar. Time of last bite.}
}
\description{
Simulate a dyad
}
