% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timegap.r
\name{timegap}
\alias{timegap}
\alias{timegap.data.frame}
\alias{timegap.matrix}
\title{Compute average time gap (time inverval) between ego-alter bites}
\usage{
timegap(data, ...)

\method{timegap}{data.frame}(data, ...)

\method{timegap}{matrix}(data, ...)
}
\arguments{
\item{data}{A data.frame or numeric matrix. The first column should be time and
the second column the ids.}

\item{...}{Further arguments passed to the method.}
}
\value{
A vector of length \code{length(unique(data[,2]))} with average time between bites
of \code{i->j} and \code{j->i}
}
\description{
Compute average time gap (time inverval) between ego-alter bites
}
