% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle.r
\name{shuffle_bites}
\alias{shuffle_bites}
\title{Shuffle bite data by sorting the intervals at individual level}
\usage{
shuffle_bites(data)
}
\arguments{
\item{data}{A two-column integer matrix. The first column holds the times,
while the second column holds the individual ids.}
}
\value{
A shuffled version of the data
}
\description{
Shuffle bite data by sorting the intervals at individual level
}
